backend/requirements.txt
flask
flask-cors
opencv-python
numpy
scipy
matplotlib
gpt4all
llama-cpp-python